3-agentcore/1_runtime/src/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore-starter-toolkit